django_2d_array/requirements.txt
asgiref==3.2.7
coverage==5.1
Django==3.0.7
django-cors-headers==3.3.0
mysqlclient==1.4.6
pytz==2020.1
sqlparse==0.3.1
